re organize trade files and folders
