+++ src/pages/Chart/Chart.tsx
@@ -45,7 +45,6 @@ import {
 import CandleChart from './Candle/CandleChart';
 import LiquidityChart from './Liquidity/LiquidityChart';
 import VolumeBarCanvas from './Volume/VolumeBarCanvas';
-import { LiquidityDataLocal } from '../Trade/TradeCharts/TradeCharts';
 import { CSSTransition } from 'react-transition-group';
 import Divider from '../../components/Global/Divider/Divider';
 import YAxisCanvas from './Axes/yAxis/YaxisCanvas';
@@ -125,6 +124,7 @@ import { filterCandleWithTransaction } from './ChartUtils/discontinuityScaleUtil
 import ChartSettings from './ChartSettings/ChartSettings';
 import useOnClickOutside, { Event } from '../../utils/hooks/useOnClickOutside';
 import ChartTooltip from './ChartTooltip/ChartTooltip';
+import { LiquidityDataLocal } from '../Trade/TradeCharts/TradeCharts/TradeCharts';
 
 interface propsIF {
     isTokenABase: boolean;
@@ -699,8 +699,7 @@ export default function Chart(props: propsIF) {
     const [bandwidth, setBandwidth] = useState(5);
     const smallScreen = useMediaQuery('(max-width: 768px)');
 
-
-    const toolbarWidth = smallScreen ? 0 :  isToolbarOpen ? 38 : 15;
+    const toolbarWidth = smallScreen ? 0 : isToolbarOpen ? 38 : 15;
 
     const [prevlastCandleTime, setPrevLastCandleTime] = useState<number>(
         lastCandleData.time,
