+++ src/pages/Chart/Liquidity/LiquiditySeries/AreaSeries.ts
@@ -1,4 +1,4 @@
-import { LiquidityDataLocal } from '../../../Trade/TradeCharts/TradeCharts';
+import { LiquidityDataLocal } from '../../../Trade/TradeCharts/TradeCharts/TradeCharts';
 import * as d3 from 'd3';
 import * as d3fc from 'd3fc';
 import { LiquidityRangeIF } from '../../../../ambient-utils/types';
