+++ src/pages/Trade/TradeCharts/TradeCharts/TradeCharts.module.css
